Kernel patch to override VM check when trying to log in with Apple ID
My Apple ID stayed logged in after the os update with only the warning that "some features are not available". However, after updating the smbios and trying to log back in with Apple ID, it returned "Verification Failed. An unknown error occurred." Sometime around Sonoma a vm check was added. Solution is kernel patch to override the check. This also re-enables bluetooth, which otherwise gets disabled through the same check.

https://forum.proxmox.com/threads/anyone-can-make-bluetooth-work-on-sonoma.153301/#post-697832